♻️ refactor(websearch): reorganize websearch module
- move headers related code to new headers.py file
- rename utils.py to filter.py to better reflect its purpose
- remove unused imports and code
- consolidate user agent generation logic in a single location
